+++ src/main/java/com/pax/market/api/sdk/java/api/base/dto/DownloadTaskDTO.java
@@ -1,0 +1,53 @@
+package com.pax.market.api.sdk.java.api.base.dto;
+
+import java.io.Serializable;
+
+public class DownloadTaskDTO implements Serializable {
+    private static final long serialVersionUID = -2112483845324750868L;
+
+    private String fileName;
+    private String downloadUrl;
+    private String signature;
+    private Long expireSeconds;
+    private String keyPairId;
+
+    public String getFileName() {
+        return fileName;
+    }
+
+    public void setFileName(String fileName) {
+        this.fileName = fileName;
+    }
+
+    public String getDownloadUrl() {
+        return downloadUrl;
+    }
+
+    public void setDownloadUrl(String downloadUrl) {
+        this.downloadUrl = downloadUrl;
+    }
+
+    public String getSignature() {
+        return signature;
+    }
+
+    public void setSignature(String signature) {
+        this.signature = signature;
+    }
+
+    public Long getExpireSeconds() {
+        return expireSeconds;
+    }
+
+    public void setExpireSeconds(Long expireSeconds) {
+        this.expireSeconds = expireSeconds;
+    }
+
+    public String getKeyPairId() {
+        return keyPairId;
+    }
+
+    public void setKeyPairId(String keyPairId) {
+        this.keyPairId = keyPairId;
+    }
+}