+++ src/main/java/com/pax/market/api/sdk/java/api/terminal/dto/TerminalLogPageResponse.java
@@ -1,0 +1,7 @@
+package com.pax.market.api.sdk.java.api.terminal.dto;
+
+import com.pax.market.api.sdk.java.api.base.dto.PageResponse;
+
+public class TerminalLogPageResponse extends PageResponse<TerminalLogDTO> {
+    private static final long serialVersionUID = -7638124013347181460L;
+}