+++ src/main/java/com/pax/market/api/sdk/java/api/terminal/dto/TerminalSystemUsageResponse.java
@@ -1,0 +1,20 @@
+/**
+ * ********************************************************************************
+ * COPYRIGHT
+ * PAX TECHNOLOGY, Inc. PROPRIETARY INFORMATION
+ * This software is supplied under the terms of a license agreement or
+ * nondisclosure agreement with PAX  Technology, Inc. and may not be copied
+ * or disclosed except in accordance with the terms in that agreement.
+ * Copyright (C) 2018 PAX Technology, Inc. All rights reserved.
+ * ********************************************************************************
+ */
+
+package com.pax.market.api.sdk.java.api.terminal.dto;
+
+import com.pax.market.api.sdk.java.api.base.dto.Response;
+
+public class TerminalSystemUsageResponse extends Response<TerminalSystemUsageDTO>{
+
+    private static final long serialVersionUID = -727612044205705882L;
+
+}