+++ src/test/java/com/pax/market/api/sdk/java/api/test/TerminalApiTest.java
@@ -12,6 +12,9 @@
 package com.pax.market.api.sdk.java.api.test;
 
 
+import com.pax.market.api.sdk.java.api.base.dto.DownloadTaskDTO;
+import com.pax.market.api.sdk.java.api.base.dto.EmptyResponse;
+import com.pax.market.api.sdk.java.api.terminal.dto.TerminalLogRequest;
 import com.pax.market.api.sdk.java.api.terminal.dto.*;
 import com.pax.market.api.sdk.java.api.terminalGroup.dto.TerminalGroupRequest;
 import com.pax.market.api.sdk.java.api.terminalGroup.dto.TerminalSnGroupRequest;
@@ -209,8 +212,8 @@ public class TerminalApiTest {
 	public void testUpdateTerminalWifiConfig(){
 		Long terminalId = 1639497804546092L;
 		TerminalWifiUpdateRequest terminalWifiUpdateRequest = new TerminalWifiUpdateRequest();
-		TerminalWifiUpdateRequest.WifiConfig wifiConfig1 = EnhancedJsonUtils.fromJson("{\"SSID\":\"pax20\",\"password\":\"12345678\",\"cipherType\":2,\"proxyType\":0}", TerminalWifiUpdateRequest.WifiConfig.class);
-		TerminalWifiUpdateRequest.WifiConfig wifiConfig2 = EnhancedJsonUtils.fromJson("{\"SSID\":\"pax30\",\"password\":\"12345678\",\"cipherType\":2,\"proxyType\":0}", TerminalWifiUpdateRequest.WifiConfig.class);
+		TerminalWifiUpdateRequest.WifiConfig wifiConfig1 = EnhancedJsonUtils.fromJson("{\"ssid\":\"pax30\",\"password\":\"12345678\",\"cipherType\":2,\"hostName\": \"1.1.1.1\",\"proxyType\":1,\"checked\": false}", TerminalWifiUpdateRequest.WifiConfig.class);
+		TerminalWifiUpdateRequest.WifiConfig wifiConfig2 = EnhancedJsonUtils.fromJson("{\"ssid\":\"pax40\",\"password\":\"12345678\",\"cipherType\":2,\"hostName\": \"1.1.1.1\",\"proxyType\":1,\"checked\": false}", TerminalWifiUpdateRequest.WifiConfig.class);
 		List<TerminalWifiUpdateRequest.WifiConfig> wifiConfigList = new ArrayList<>();
 		wifiConfigList.add(wifiConfig1);
 		wifiConfigList.add(wifiConfig2);
@@ -289,87 +292,104 @@ public class TerminalApiTest {
 		Assert.assertTrue(result.getBusinessCode() == 0);
 	}
 
-//	===============20240507新增
-@Test
-public void testGetTerminalIncludeAccessoryInfoBySn() {
-	String serialNo = "SN6132522";
-	Result<TerminalDTO> getResult = terminalApi.getTerminalBySn(serialNo,true);
-	logger.debug("Result of get terminal: {}",getResult.toString());
-	Assert.assertTrue(getResult.getBusinessCode() == 0);
-}
+	private final String serialNo = "TESTBYSN";
+	private final Long terminalId = 1620067374596136L;
+	private final String factoryName = "PAX";
+	private final String modelName = "E500";
+
+	@Test
+	public void testTerminalBySn() {
+		testUpdateTerminalBySn();
+		testActivateTerminalBySn();
+		testGetTerminalBySn();
+		testCopyTerminalBySn();
+
+		testSendTerminalMessage();
+		testBatchAddTerminalToGroupBySn();
+		testUpdateTerminalConfigBySn();
+		testMoveTerminalBySn();
+		testPushTerminalActionBySn();
+
+		testGetTerminalIncludeAccessoryInfoBySn();
+		testGetTerminalIncludeInstalledApksBySn();
+		testGetTerminalConfigBySn();
+		testGetTerminalPedBySn();
+
+		testDisableTerminalBySn();
+		testDeleteTerminalBySn();
+	}
+	@Test
+	public void testGetTerminalIncludeAccessoryInfoBySn() {
+		Result<TerminalDTO> getResult = terminalApi.getTerminalBySn(serialNo,true);
+		logger.debug("Result of get terminal: {}",getResult.toString());
+        Assert.assertEquals(0, getResult.getBusinessCode());
+	}
 
 	@Test
 	public void testGetTerminalIncludeInstalledApksBySn() {
-		String serialNo = "SN9989311";
 		Result<TerminalDTO> getResult = terminalApi.getTerminalBySn(serialNo,false, true);
 		logger.debug("Result of get terminal: {}",getResult.toString());
-		Assert.assertTrue(getResult.getBusinessCode() == 0);
+        Assert.assertEquals(0, getResult.getBusinessCode());
 	}
 
 	@Test
 	public void testBatchAddTerminalToGroupBySn(){
 		TerminalSnGroupRequest groupRequest = new TerminalSnGroupRequest();
 		Set<String> serialNoList = new HashSet<>();
-		serialNoList.add("SN9989311");
+		serialNoList.add(serialNo);
 //		serialNoList.add("QWJQW028");
 		Set<Long> groupIds = new HashSet<>();
-		groupIds.add(1001567300L);
+		groupIds.add(16626L);
 		groupRequest.setSerialNos(serialNoList);
 		groupRequest.setGroupIds(groupIds);
 		Result<String> result = terminalApi.batchAddTerminalToGroupBySn(groupRequest);
 		logger.debug("Result of search terminal: {}",result.toString());
-		Assert.assertTrue(result.getBusinessCode() == 0);
+        Assert.assertEquals(0, result.getBusinessCode());
 
 	}
 
 	@Test
 	public void testUpdateTerminalConfigBySn(){
-		String serialNo = "SN9989311";
 		TerminalReplacementUpdateRequest terminalConfigUpdateRequest = new TerminalReplacementUpdateRequest();
 		terminalConfigUpdateRequest.setAllowReplacement(true);
 		Result<String> result = terminalApi.updateTerminalConfigBySn(serialNo,terminalConfigUpdateRequest);
 		logger.debug("Result of update Terminal Config: {}",result.toString());
-		Assert.assertTrue(result.getBusinessCode() == 0);
+        Assert.assertEquals(0, result.getBusinessCode());
 
 	}
 
 	@Test
 	public void testGetTerminalConfigBySn(){
-		String serialNo = "SN9989311";
 		Result<TerminalConfigDTO> result = terminalApi.getTerminalConfigBySn(serialNo);
 		logger.debug("Result of get Terminal Config: {}",result.toString());
-		Assert.assertTrue(result.getBusinessCode() == 0);
+        Assert.assertEquals(0, result.getBusinessCode());
 
 	}
 
 
 	@Test
 	public void testGetTerminalPedBySn(){
-		String serialNo = "SN9989311";
 		Result<TerminalPedDTO> result = terminalApi.getTerminalPedBySn(serialNo);
 		logger.debug("Result of get Terminal ped: {}",result.toString());
-		Assert.assertTrue(result.getBusinessCode() == 0);
+        Assert.assertEquals(0, result.getBusinessCode());
 	}
 
 	@Test
 	public void testMoveTerminalBySn() {
-		String serialNo = "SN9989311";
-		Result<String> result = terminalApi.moveTerminalBySn(serialNo, "PAX", "6666");
+		Result<String> result = terminalApi.moveTerminalBySn(serialNo, "shifan", "merchant0");
 		logger.debug("Result of move Terminal {}",result.toString());
 		Assert.assertTrue(result.getBusinessCode() == 0);
 	}
 
 	@Test
 	public void testPushTerminalActionBySn() {
-		String serialNo = "SN6132522";
-		Result<String> result = terminalApi.pushCmdToTerminalBySn(serialNo, TerminalApi.TerminalPushCmd.Unlock);
+		Result<String> result = terminalApi.pushCmdToTerminalBySn(serialNo, TerminalApi.TerminalPushCmd.Lock);
 		logger.debug("Result of push terminal action {}",result.toString());
 		Assert.assertEquals(0, result.getBusinessCode());
 	}
 
 	@Test
 	public void testActivateTerminalBySn() {
-		String serialNo = "08200001522336";
 		Result<String> result = terminalApi.activateTerminalBySn(serialNo);
 		logger.debug("Result of activateTerminalBySn {}",result.toString());
 		Assert.assertEquals(0, result.getBusinessCode());
@@ -377,7 +397,6 @@ public void testGetTerminalIncludeAccessoryInfoBySn() {
 
 	@Test
 	public void testDeleteTerminalBySn() {
-		String serialNo = "08200001522336";
 		Result<String> result = terminalApi.deleteTerminalBySn(serialNo);
 		logger.debug("Result of deleteTerminalBySn {}",result.toString());
 		Assert.assertEquals(0, result.getBusinessCode());
@@ -385,11 +404,9 @@ public void testGetTerminalIncludeAccessoryInfoBySn() {
 
 	@Test
 	public void testCopyTerminalBySn() {
-//		String serialNo = "EMPTY1716345320431";
-		String serialNo = "08200001522336";
 		TerminalSnCopyRequest request=new TerminalSnCopyRequest();
 		request.setSourceSerialNo(serialNo);
-		request.setSerialNo("");
+		request.setSerialNo("TESTCOPYBYSN");
 		request.setStatus(TerminalStatus.Inactive);
 		request.setName("MyTerminal240524");
 		Result<TerminalDTO> result = terminalApi.copyTerminalBySn(request);
@@ -400,37 +417,120 @@ public void testGetTerminalIncludeAccessoryInfoBySn() {
 	@Test
 	public void testUpdateTerminalBySn() {
 		TerminalUpdateRequest updateRequest = new TerminalUpdateRequest();
-		updateRequest.setSerialNo("SN9989311");
+		updateRequest.setSerialNo(serialNo);
 		updateRequest.setName("Terminal No.7637");
-		updateRequest.setResellerName("PAX");
+		updateRequest.setResellerName("shifan");
 		updateRequest.setRemark("new Remark2024");
 		updateRequest.setModelName("A920");
+		updateRequest.setMerchantName("merchant to update");
 
-		Result<TerminalDTO> terminalDTOResult = terminalApi.updateTerminalBySn("SN9989311", updateRequest);
-		if(terminalDTOResult.getData()!=null) {
-			logger.debug("Result of testCopyTerminalBySn {}",terminalDTOResult.getData().toString());
-		}else{
-			logger.debug("Result of testCopyTerminalBySn {}", terminalDTOResult.toString());
-		}
+		Result<TerminalDTO> terminalDTOResult = terminalApi.updateTerminalBySn(serialNo, updateRequest);
+		logger.debug("Result of testCopyTerminalBySn {}", terminalDTOResult.toString());
 		Assert.assertEquals(0, terminalDTOResult.getBusinessCode());
 	}
 
 	@Test
 	public void testDisableTerminalBySn() {
-		Result<String> result = terminalApi.disableTerminalBySn("SN9989311");
+		Result<String> result = terminalApi.disableTerminalBySn(serialNo);
 		logger.debug("Result of testDisableTerminalBySn {}",result.toString());
 		Assert.assertEquals(0, result.getBusinessCode());
 	}
 
 	@Test
 	public void testGetTerminalBySn() {
-		Result<TerminalDTO> terminalBySn = terminalApi.getTerminalBySn("SN9989311");
-		if(terminalBySn.getData()!=null) {
-			logger.debug("Result of testGetTerminalBySn {}",terminalBySn.getData().toString());
-		}else{
-			logger.debug("Result of testGetTerminalBySn {}", terminalBySn.toString());
-		}
+		Result<TerminalDTO> terminalBySn = terminalApi.getTerminalBySn(serialNo);
+		logger.debug("Result of testGetTerminalBySn {}", terminalBySn.toString());
 		Assert.assertEquals(0, terminalBySn.getBusinessCode());
 	}
 
+	@Test
+	public void testSendTerminalMessage() {
+		//terminal is online
+		TerminalMessageRequest request = new TerminalMessageRequest();
+		request.setTitle("Test Title");
+		request.setContent("Test Content");
+		Result<String> result = terminalApi.pushTerminalMessage(terminalId,request);
+		logger.debug("Result of get result: {}",result.toString());
+		Assert.assertEquals(0, result.getBusinessCode());
+
+		Result<String> resultBySn = terminalApi.pushTerminalMessageBySn(serialNo,request);
+		logger.debug("Result of get result: {}",resultBySn.toString());
+		Assert.assertEquals(0, resultBySn.getBusinessCode());
+	}
+
+	@Test
+	public void testTerminalSystemUsageBySN() {
+		Result<TerminalSystemUsageDTO> terminalSystemUsage = terminalApi.getTerminalSystemUsageBySn("SUBSN108");
+		if(terminalSystemUsage.getData()!=null) {
+			logger.debug("Result of testTerminalCpuStatistic {}",terminalSystemUsage.getData().toString());
+		}else{
+			logger.debug("Result of testTerminalCpuStatistic {}", terminalSystemUsage.toString());
+		}
+		Assert.assertEquals(0, terminalSystemUsage.getBusinessCode());
+	}
+
+	@Test
+	public void testTerminalSystemUsageById() {
+		Result<TerminalSystemUsageDTO> terminalSystemUsage = terminalApi.getTerminalSystemUsageById(1639491781525547L);
+		if(terminalSystemUsage.getData()!=null) {
+			logger.debug("Result of testTerminalCpuStatistic {}",terminalSystemUsage.getData().toString());
+		}else{
+			logger.debug("Result of testTerminalCpuStatistic {}", terminalSystemUsage.toString());
+		}
+		Assert.assertEquals(0, terminalSystemUsage.getBusinessCode());
+	}
+
+	@Test
+	public void testCollectTerminalLog() {
+		TerminalLogRequest request = new TerminalLogRequest();
+		request.setType("L");
+		Result<String> result = terminalApi.collectTerminalLog(terminalId,request);
+		Assert.assertEquals(0, result.getBusinessCode());
+	}
+
+	@Test
+	public void testCollectTerminalLogBySn() {
+		TerminalLogRequest request = new TerminalLogRequest();
+		request.setType("L");
+		Result<String> result = terminalApi.collectTerminalLogBySn(serialNo,request);
+		Assert.assertEquals(0, result.getBusinessCode());
+	}
+
+	@Test
+	public void testSearchTerminalLog() {
+		Result<TerminalLogDTO> result = terminalApi.searchTerminalLog(1, 10, terminalId);
+		Assert.assertEquals(0, result.getBusinessCode());
+	}
+
+	@Test
+	public void testSearchTerminalLogBySn() {
+		Result<TerminalLogDTO> result = terminalApi.searchTerminalLogBySn(1, 10, serialNo);
+		Assert.assertEquals(0, result.getBusinessCode());
+	}
+
+	@Test
+	public void testGetTerminalLogDownloadTask() {
+		Long terminalLogId = 1018664888L;
+		Result<DownloadTaskDTO> result = terminalApi.getTerminalLogDownloadTask(terminalId, terminalLogId);
+		Assert.assertEquals(0, result.getBusinessCode());
+	}
+
+	@Test
+	public void testGetTerminalLogDownloadTaskBySn() {
+		Long terminalLogId = 1018664888L;
+		Result<DownloadTaskDTO> result = terminalApi.getTerminalLogDownloadTaskBySn(serialNo, terminalLogId);
+		Assert.assertEquals(0, result.getBusinessCode());
+	}
+
+	@Test
+	public void testChangeModelBySN() {
+		Result<EmptyResponse> responseResult = terminalApi.changeModelBySN("SUBSN108", factoryName, modelName);
+		Assert.assertEquals(0, responseResult.getBusinessCode());
+	}
+
+	@Test
+	public void testChangeModelById() {
+		Result<EmptyResponse> responseResult = terminalApi.changeModel(1639491781525547L, factoryName, modelName);
+		Assert.assertEquals(0, responseResult.getBusinessCode());
+	}
 }